requirements.txt
# Core project
requests
schedule
python-dotenv
filelock

# SpiderFoot critical deps (FIXED)
cryptography==3.4.8
pyOpenSSL==21.0.0
dnspython==2.3.0
cherrypy==18.8.0
cherrypy-cors==1.6.0

# SpiderFoot full dependencies
adblockparser>=0.7,<1
ExifRead>=2.3.2,<3
Mako>=1.2.4,<2
beautifulsoup4>=4.11.2,<5
lxml>=4.9.2,<5
netaddr>=0.8.0,<1
pysocks>=1.7.1,<2
ipwhois>=1.1.0,<1.2.0
ipaddr>=2.2.0,<3
phonenumbers>=8.13.6,<9
pygexf>=0.2.2,<0.3
PyPDF2>=1.28.6,<2
python-whois>=0.7.3,<0.8
secure>=0.3.0,<0.4.0
python-docx>=0.8.11,<0.9
python-pptx>=0.6.21,<0.7
networkx>=2.6.3,<2.7
publicsuffixlist>=0.9.3,<0.10
openpyxl>=3.1.1,<4
pyyaml>=6.0.0,<7